+++ src/models/pending-refresh-tokens/index.ts
@@ -1,4 +1,11 @@
+import _ from "lodash";
 import { redis } from "@/common/redis";
+
+export type RefreshTokens = {
+  collection: string;
+  contract: string;
+  tokenId: string;
+};
 
 /**
  * Class that manage redis list of tokens, pending metadata refresh
@@ -10,15 +17,26 @@ export class PendingRefreshTokens {
     this.key += `:${method}`;
   }
 
-  public async add(token: string[], prioritized = false) {
+  public async add(refreshToken: RefreshTokens[], prioritized = false) {
     if (prioritized) {
-      return await redis.lpush(this.key, token);
+      return await redis.lpush(
+        this.key,
+        _.map(refreshToken, (token) => JSON.stringify(token))
+      );
     } else {
-      return await redis.rpush(this.key, token);
+      return await redis.rpush(
+        this.key,
+        _.map(refreshToken, (token) => JSON.stringify(token))
+      );
     }
   }
 
-  public async get(count = 20) {
-    return await redis.lpop(this.key, count);
+  public async get(count = 20): Promise<RefreshTokens[]> {
+    const refreshTokens = await redis.lpop(this.key, count);
+    if (refreshTokens) {
+      return _.map(refreshTokens, (refreshToken) => JSON.parse(refreshToken) as RefreshTokens);
+    }
+
+    return [];
   }
 }